refactor get_all_zuul_yaml_files() function
get_all_zuul_yaml_files() now returns a dictionary where
the keys contain the extensions of files found:
    - extension: .yaml ---> key: "good_yaml"
    - extension: .yml ---> key: "bad_yaml"
for each key(file extension) there is a list of paths
where files with the respective extension can be found.
test_utils.py was modified to account for the changes mentioned
above and all tests are passing.